+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2021-02-23
+
+### Added
+
+-   additional binanceClient.fetchBalance() data instead of config variables
+
+### Removed
+
+-   a bunch junk no longerd needed
+
 ## [0.5.1] - 2021-02-22
 
 ### Added
@@ -130,6 +140,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Working demo and API connections
 
+[0.5.2]: https://github.com/JesseChamberlain/binance-trader/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/JesseChamberlain/binance-trader/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/JesseChamberlain/binance-trader/compare/v0.4.1...v0.4.2